Also move headers into their own directory, and test_hmd_driver into its own
