+++ headers/osvr_device_properties.h
