+++ headers/osvr_tracked_device.h
