+++ src/ServerDriver_OSVR.cpp
@@ -22,7 +22,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <ServerDriver_OSVR.h>
+#include "headers/ServerDriver_OSVR.h"
 
 vr::EVRInitError ServerDriver_OSVR::Init(vr::IDriverLog* driver_log, vr::IServerDriverHost* driver_host, const char* user_driver_config_dir, const char* driver_install_dir)
 {
